refactor v1 trace collection helpers
